[Documentation] Add advanced tutorial (#146)
* add advanced tutorial

* change notebook python run

* tests of methods embedding and pulse

* coloring solver with abstract methods

* [Fix] For pulse shaping, we need to use the qubit positions by index … (#140)

After in-depth debugging, it turns out that we were shaping the pulse based on an order of qubits
that made no sense. This change fixes the issue, but we still need a good test to avoid any
regression.

* [Fix] Isolated node removal doesn't eat solutions anymore - resolves … (#144)

While there may be other cases in which the preprocessor eats solutions,
so far, this is the only one we managed to detect.

* [Doc] Typo fix in tutorial 1 (#148)

* [Feature] Weighted detuning (#131)

This is the last missing part for getting WMIS to work.

* add detuning

* fix tuto with rebase of detuning

---------

Co-authored-by: David Teller <D.O.Teller@gmail.com>